Update README.md
This work was aimed at investigating, how non-destructively measured modulus of elasticity, specific gravity, moisture content, together with the information on wood type, could be used for prediction of various mechanical properties of wood, e.g. bending, compressive, shear, and tensile strength (parallel and perpendicular to the fiber direction), hardness and impact bending resistance. To that end, several machine learning approaches, namely, linear regression, partial least squares regression, random forest regression, support vector regression and support vector regression combined with a regressor chain, were exploited. The evaluation of the methods' performance and selection of the best model among the employed models was based on the values of the coefficient of determination (${R^2}$) from the regression of the predicted values on the observed ones using the test data, and the test root mean squared error ($RMSE$). Accordingly, support vector regression combined with a regressor chain, $SVR(RC)$, capable of effectively learning non-linear relationships in the data and considering the potential correlations among different targets, showed superior performance in predicting the mechanical properties of wood. Additional assessment of the model's accuracy involved visual inspection of the shapes of the predicted and observed distribution functions for the said properties. It was observed that the performance of the trained $SVR(RC)$ model differed among the properties, being high for the properties parallel to the fiber direction; e.g. compressive strength and side hardness, while being substantially lower for work to maximum load.